requirements.txt
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
arrow==1.2.3
attrs==21.4.0
cairocffi==1.4.0
CairoSVG==2.6.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
colour==0.1.5
cpe==1.2.1
cssselect2==0.7.0
cybox==2.1.0.21
deepdiff==6.2.3
defusedxml==0.7.1
drawSvg==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fqdn==1.5.1
idna==3.4
imageio==2.25.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
isoduration==20.11.0
jsonpointer==2.3
jsonschema==4.17.3
loguru==0.6.0
lxml==4.9.2
maec==4.1.0.17
Markdown==3.4.1
markdown-it-py==2.1.0
mdurl==0.1.2
mitreattack-python==2.0.3
mixbox==1.0.5
netaddr==0.8.0
numpy==1.24.1
openpyxl==3.0.10
ordered-set==4.1.0
orjson==3.8.5
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
pluralizer==1.2.0
pooch==1.6.0
pycountry==22.3.5
pycparser==2.21
Pygments==2.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-cache==0.9.8
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.0
simplejson==3.18.1
six==1.16.0
stix==1.2.0.11
stix2==3.0.1
stix2-elevator==4.1.7
stix2-patterns==2.0.0
stix2-validator==3.1.2
stixmarx==1.0.8
tabulate==0.9.0
taxii2-client==2.3.0
tinycss2==1.2.1
tqdm==4.64.1
typer==0.7.0
typing-extensions==4.4.0
uri-template==1.2.0
url-normalize==1.4.3
urllib3==1.26.14
weakrefmethod==1.0.3
webcolors==1.12
webencodings==0.5.1
XlsxWriter==3.0.7
zipp==3.11.0
